% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{package}
\name{s4wrapper-package}
\alias{s4wrapper}
\alias{s4wrapper-package}
\title{s4wrapper: Community-Provided Methods and Extensions for the Seurat Object}
\description{
s4wrapper is a collection of community-provided methods and extensions for Seurat, and are able to be updated much more frequently.
}
\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Zan Yuan \email{seqyuan@gmail.com} (\href{https://orcid.org/0000-0001-5582-5120}{ORCID})
}

}
